last commit

+++ transtu.tn/src/main/java/com/pfe/Service/ItineraireService.java
@@ -25,23 +25,47 @@ public class ItineraireService {
     private PointGPSRepository pointGPSRepository;
 
     public Itineraire createItineraire(ItineraireRequest request) {
+
+        Optional<Itineraire> existingItineraire = itineraireRepository.findByCode(request.getCode());
+        if (existingItineraire.isPresent()) {
+            throw new RuntimeException("Un itinéraire avec le même code existe déjà");
+        }
+
         Itineraire itineraire = convertRequestToItineraire(request);
         return itineraireRepository.save(itineraire);
     }
-
     public Itineraire updateItineraire(Long id, ItineraireRequest request) {
         Optional<Itineraire> existingItineraire = itineraireRepository.findById(id);
         if (existingItineraire.isPresent()) {
-            Itineraire itineraire = convertRequestToItineraire(request);
-            itineraire.setId(id);
-            return itineraireRepository.save(itineraire);
+            Itineraire itineraire = existingItineraire.get();
+
+            boolean isUpdated = false;
+            if (!itineraire.getCode().equals(request.getCode())) {
+                itineraire.setCode(request.getCode());
+                isUpdated = true;
+            }
+            if (!itineraire.getDescription().equals(request.getDescription())) {
+                itineraire.setDescription(request.getDescription());
+                isUpdated = true;
+            }
+
+            if (isUpdated) {
+                return itineraireRepository.save(itineraire);
+            } else {
+                return itineraire;
+            }
+        } else {
+            throw new IllegalArgumentException("Itinéraire non trouvé avec l'ID : " + id);
+        }
+    }
+    public void deleteItineraire(Long id) {
+   
+        Optional<Itineraire> existingItineraire = itineraireRepository.findById(id);
+        if (existingItineraire.isPresent()) {
+            itineraireRepository.deleteById(id);
         } else {
             throw new RuntimeException("Itinéraire non trouvé avec l'ID : " + id);
         }
-    }
-
-    public void deleteItineraire(Long id) {
-        itineraireRepository.deleteById(id);
     }
 
     private Itineraire convertRequestToItineraire(ItineraireRequest request) {
@@ -67,4 +91,5 @@ public class ItineraireService {
     public List<Itineraire> getAllItineraires() {
         return itineraireRepository.findAll();
     }
+    
     }
